requirements.txt
streamlit
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
pypdf
python-dotenv
gTTS